requirement.txt
flask-2.2.2
Jinja2-3.1.2 
MarkupSafe-2.1.1 
Werkzeug-2.2.2 
click-8.1.3 
colorama-0.4.6 
importlib-metadata-5.1.0 
itsdangerous-2.1.2 
zipp-3.11.0
SQLAlchemy-1.4.44 
flask_sqlalchemy-3.0.2 
greenlet-2.0.1
flask_marshmallow-0.14.0 
marshmallow-3.19.0 
packaging-21.3 
pyparsing-3.0.9 
six-1.16.0
flask_httpauth-4.7.0
boto3-1.26.20 
botocore-1.29.20 
jmespath-1.0.1 
python-dateutil-2.8.2 
s3transfer-0.6.0 
urllib3-1.26.13
aniso8601-9.0.1 
attrs-22.1.0 
flask_restplus-0.13.0 
jsonschema-4.17.3 
pyrsistent-0.19.2 
pytz-2022.6
marshmallow-sqlalchemy-0.28.1
healthcheck-1.3.3
mysqlclient-2.1.1
certifi-2022.9.24 
charset-normalizer-2.1.1 
idna-3.4 
requests-2.28.1
PyYAML-6.0 
flasgger-0.9.5 
mistune-2.0.4